chore: move GrpcConnector to common package
